requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pymysql==1.1.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-dotenv==1.0.1
pydantic==2.10.3
